eradicate `util`
