+++ collectors/modality-probe-collector-common/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "modality-probe-collector-common"
+version = "0.1.0"
+authors = ["dan pittman <dan@auxon.io>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+chrono = { version = "0.4", features = ["serde"] }
+csv = "1.1"
+
+modality-probe = { path = "../../", features = ["std"] }
+
+[dev-dependencies]
+proptest = "0.9"
