+++ collectors/modality-probe-udp-collector/src/completions.rs
